+++ pysat/tests/test_files.py
@@ -49,7 +49,7 @@ def create_files(inst, start, stop, freq=None, use_doy=True, root_fname=None, co
 
     if freq is None:
         freq = '1D'
-    dates = pysat.utils.time.season_date_range(start, stop, freq=freq)
+    dates = pysat.utils.time.create_date_range(start, stop, freq=freq)
 
     if root_fname is None:
         root_fname = ''.join(('pysat_testing_junk_{year:04d}_gold_{day:03d}_',
@@ -336,7 +336,7 @@ class TestBasics():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False, root_fname=root_fname)
         # create the same range of dates
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         pysat.instruments.pysat_testing.list_files = list_files
         inst = pysat.Instrument(platform='pysat', name='testing',
                                 update_files=True)
@@ -412,7 +412,7 @@ class TestInstrumentWithFiles():
                      use_doy=False,
                      root_fname=self.root_fname)
         start = pysat.datetime(2007, 12, 31)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         assert (np.all(self.testInst.files.files.index == dates))
 
@@ -460,7 +460,7 @@ class TestInstrumentWithFiles():
     def test_refresh_on_unchanged_files(self):
         start = pysat.datetime(2007, 12, 31)
         stop = pysat.datetime(2008, 1, 10)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         assert (np.all(self.testInst.files.files.index == dates))
 
@@ -472,7 +472,7 @@ class TestInstrumentWithFiles():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False,
                      root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         new_files = self.testInst.files.get_new()
 
@@ -486,7 +486,7 @@ class TestInstrumentWithFiles():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False,
                      root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         self.testInst.files.refresh()
         self.testInst.files.refresh()
@@ -501,7 +501,7 @@ class TestInstrumentWithFiles():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False,
                      root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files = self.testInst.files.get_new()
         assert (np.all(new_files.index == dates))
 
@@ -513,7 +513,7 @@ class TestInstrumentWithFiles():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False,
                      root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files = self.testInst.files.get_new()
 
         start = pysat.datetime(2008, 1, 15)
@@ -522,7 +522,7 @@ class TestInstrumentWithFiles():
         create_files(self.testInst, start, stop, freq='100min',
                      use_doy=False,
                      root_fname=self.root_fname)
-        dates2 = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates2 = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files2 = self.testInst.files.get_new()
         assert (np.all(new_files.index == dates) &
                 np.all(new_files2.index == dates2))
@@ -531,7 +531,7 @@ class TestInstrumentWithFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 12)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
 
         # remove files, same number as will be added
         to_be_removed = len(dates)
@@ -556,7 +556,7 @@ class TestInstrumentWithFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 15)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
 
         self.testInst = \
             pysat.Instrument(inst_module=pysat.instruments.pysat_testing,
@@ -589,7 +589,7 @@ class TestInstrumentWithFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 15)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='1D')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='1D')
 
         # clear out old files, create new ones
         remove_files(self.testInst)
@@ -650,7 +650,7 @@ def create_versioned_files(inst, start=None, stop=None, freq='1D',
         start = pysat.datetime(2009, 1, 1)
     if stop is None:
         stop = pysat.datetime(2013, 12, 31)
-    dates = pysat.utils.time.season_date_range(start, stop, freq=freq)
+    dates = pysat.utils.time.create_date_range(start, stop, freq=freq)
 
     versions = np.array([1, 2])
     revisions = np.array([0, 1])
@@ -765,7 +765,7 @@ class TestInstrumentWithVersionedFiles():
                                root_fname=self.root_fname)
         # create list of dates for all files that should be there
         start = pysat.datetime(2007, 12, 31)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         # update instrument file list
         self.testInst.files.refresh()
         assert (np.all(self.testInst.files.files.index == dates))
@@ -774,7 +774,7 @@ class TestInstrumentWithVersionedFiles():
 
         start = pysat.datetime(2007, 12, 31)
         stop = pysat.datetime(2008, 1, 10)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         assert (np.all(self.testInst.files.files.index == dates))
 
@@ -786,7 +786,7 @@ class TestInstrumentWithVersionedFiles():
         create_versioned_files(self.testInst, start, stop, freq='100min',
                                use_doy=False,
                                root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         new_files = self.testInst.files.get_new()
 
@@ -800,7 +800,7 @@ class TestInstrumentWithVersionedFiles():
         create_versioned_files(self.testInst, start, stop, freq='100min',
                                use_doy=False,
                                root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         self.testInst.files.refresh()
         self.testInst.files.refresh()
         self.testInst.files.refresh()
@@ -816,7 +816,7 @@ class TestInstrumentWithVersionedFiles():
         create_versioned_files(self.testInst, start, stop, freq='100min',
                                use_doy=False,
                                root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files = self.testInst.files.get_new()
         assert (np.all(new_files.index == dates))
 
@@ -828,7 +828,7 @@ class TestInstrumentWithVersionedFiles():
         create_versioned_files(self.testInst, start, stop, freq='100min',
                                use_doy=False,
                                root_fname=self.root_fname)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files = self.testInst.files.get_new()
 
         start = pysat.datetime(2008, 1, 15)
@@ -837,7 +837,7 @@ class TestInstrumentWithVersionedFiles():
         create_versioned_files(self.testInst, start, stop, freq='100min',
                                use_doy=False,
                                root_fname=self.root_fname)
-        dates2 = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates2 = pysat.utils.time.create_date_range(start, stop, freq='100min')
         new_files2 = self.testInst.files.get_new()
         assert (np.all(new_files.index == dates) &
                 np.all(new_files2.index == dates2))
@@ -846,7 +846,7 @@ class TestInstrumentWithVersionedFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 12)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         # remove files, same number as will be added
         to_be_removed = len(dates)
         for the_file in os.listdir(self.testInst.files.data_path):
@@ -874,7 +874,7 @@ class TestInstrumentWithVersionedFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 15)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='100min')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='100min')
         pysat.instruments.pysat_testing.list_files = list_versioned_files
         self.testInst = \
             pysat.Instrument(inst_module=pysat.instruments.pysat_testing,
@@ -906,7 +906,7 @@ class TestInstrumentWithVersionedFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 15)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='1D')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='1D')
 
         # clear out old files, create new ones
         remove_files(self.testInst)
@@ -935,7 +935,7 @@ class TestInstrumentWithVersionedFiles():
         # create new files and make sure that new files are captured
         start = pysat.datetime(2008, 1, 11)
         stop = pysat.datetime(2008, 1, 15)
-        dates = pysat.utils.time.season_date_range(start, stop, freq='1D')
+        dates = pysat.utils.time.create_date_range(start, stop, freq='1D')
 
         # clear out old files, create new ones
         remove_files(self.testInst)