+++ pysat/utils/time.py
@@ -124,7 +124,7 @@ def calc_freq(index):
     return freq
 
 
-def season_date_range(start, stop, freq='D'):
+def create_date_range(start, stop, freq='D'):
     """
     Return array of datetime objects using input frequency from start to stop