+++ CHANGELOG.md
@@ -2,14 +2,30 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.X.X] - 2019-07-17
+## [2.X.X] - 2019-09-16
 - New Features
    - Added new velocity format options to utils.coords.scale_units
    - Improved failure messages for utils.coords.scale_units
    - Added some tests for model_utils
+   - Added option to to_netCDF that names variables in the written file
+   based upon the strings in the Instrument.meta object
+   - Improved compatibility with NASA ICON's file standards
+   - Improved file downloading for Kp
+- Code Restructure
+  - Move `computational_form` to `ssnl`
+  - Move `scale_units` to `utils._core`
+  - Rename `season_date_range` as `create_date_range`
+  - Added deprecation warnings to stat functions
 - Bug fix
    - Fixed implementation of utils routines in model_utils and jro_isr
    - Fixed error catching bug in model_utils
+   - Updated Instrument.concat_data for consistency across pandas and xarray. Includes support for user provided keywords.
+   - Fixed error introduced by upstream change in NOAA F10.7 file format
+   - Fixed issue with data access via Instrument object using time and name slicing and xarray. Added unit test.
+   - Updated travis.yml to work under pysat organization
+   - Added missing requirements (matplotlib, netCDF4)
+- Documentation
+  - Added info on how to cite the code and package.
 
 ## [2.0.0] - 2019-07-11
  - New Features